+++ src/ASPNETHandler/HTTPServiceContext.cs
@@ -1,0 +1,25 @@
+﻿/*----------------------------------------------------------
+This Source Code Form is subject to the terms of the 
+Mozilla Public License, v.2.0. If a copy of the MPL 
+was not distributed with this file, You can obtain one 
+at http://mozilla.org/MPL/2.0/.
+----------------------------------------------------------*/
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using ScriptEngine.Machine;
+using ScriptEngine.Machine.Contexts;
+using ScriptEngine.HostedScript.Library;
+using ScriptEngine.HostedScript.Library.Binary;
+
+namespace ScriptEngine.HostedScript.Library.HTTPService
+{
+    [ContextClass("HTTPСервисКонтекст", "HTTPServiceContext")]
+    class HTTPServiceContextImpl
+    {
+    }
+}
